requirements.txt
appdirs==1.4.3
aspy.yaml==1.3.0
cfgv==3.0.0
cycler==0.10.0
distlib==0.3.0
filelock==3.0.12
identify==1.4.11
importlib-metadata==1.5.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.3
nodeenv==1.3.5
numpy==1.18.1
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
pre-commit==2.0.1
py==1.8.1
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
toml==0.10.0
virtualenv==20.0.4
yapf==0.29.0
zipp==2.2.0
